refactor: Merge in ErrorOverlay too
