+++ src/lib/repl.js
@@ -1,2 +1,3 @@
 export { default as CodeEditor } from '../components/code-editor';
 export { default as Runner } from '../components/controllers/repl/runner';
+export { ErrorOverlay } from '../components/controllers/repl/error-overlay';
